+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release Notes for Craft CMS 3.x
 
+## Unreleased
+- Fixed an error that could occur when processing template caches in a console request, if a globally-scoped template cache was processed before it.
+
 ## 3.7.62 - 2022-12-13
 
 - Fixed a bug where it wasn’t possible to enter `0` in Number fields and numeric column cells within editable tables, using certain keyboard layouts. ([#12412](https://github.com/craftcms/cms/issues/12412))